Fix Terraform Plan for Docker Repository (#7)
Now that we have Terraform CI in place, we can fix up the resources so that our Docker Repository is created. This also means we should make objects that can refer to the object representation of the project, such that Terraform will not try and create objects within a project without first creating the project itself.